refactor : 인수테스트 리팩토링 - 패키지 변경 - 추상클래스 생성
